requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-multipart
pillow
numpy
torch
torchvision
timm
python-dotenv
pillow-heif
python-magic
slowapi
aiofiles
cloudinary